improve to_seed
